Make timeout an integer.

+++ java-jet/src/main/java/com/hbm/devices/jet/JetConnection.java
@@ -1,5 +1,5 @@
 package com.hbm.devices.jet;
 
 public interface JetConnection {
-	void connect(double timeoutMs);
+	void connect(int timeoutMs);
 }
